+++ integrations/aspire/playground/Scalar.Aspire.BookService/Program.cs
@@ -1,0 +1,10 @@
+var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddOpenApi();
+var app = builder.Build();
+
+app.MapGet("/", () => "Hello World!");
+
+app.MapOpenApi("/swagger/{documentName}.json");
+
+app.Run();