+++ integrations/aspnetcore/src/Scalar.AspNetCore/Options/ScalarOptions.cs
@@ -64,6 +64,7 @@ public sealed class ScalarOptions
     /// Gets or sets whether to hide the "Download OpenAPI Specification" button.
     /// </summary>
     /// <value>The default value is <c>false</c>.</value>
+    [Obsolete($"This property is obsolete and will be removed in a future release. Please use the '{nameof(DocumentDownloadType)}' property instead.")]
     public bool HideDownloadButton { get; set; }
 
     /// <summary>
@@ -280,4 +281,10 @@ public sealed class ScalarOptions
     /// When set to <c>true</c>, the authentication state will be stored in the browser's local storage and restored when the user returns to the page.
     /// </remarks>
     public bool PersistentAuthentication { get; set; }
+
+    /// <summary>
+    /// Gets or sets the type of document download available for the API documentation.
+    /// </summary>
+    /// <value>The default value is <c>null</c>.</value>
+    public DocumentDownloadType? DocumentDownloadType { get; set; }
 }