+++ packages/api-reference/src/components/ApiReferenceLayout.vue
@@ -43,12 +43,11 @@ import ClassicHeader from '@/components/ClassicHeader.vue'
 import { Content } from '@/components/Content'
 import GettingStarted from '@/components/GettingStarted.vue'
 import MobileHeader from '@/components/MobileHeader.vue'
-import { Sidebar } from '@/components/Sidebar'
-import { ApiClientModal } from '@/features/ApiClientModal'
-import { useDocumentSource } from '@/features/DocumentSource'
-import { OPENAPI_VERSION_SYMBOL } from '@/features/DownloadLink'
+import { ApiClientModal } from '@/features/api-client-modal'
+import { useDocumentSource } from '@/features/document-source'
+import { OPENAPI_VERSION_SYMBOL } from '@/features/download-link'
 import { SearchButton } from '@/features/Search'
-import { useSidebar } from '@/features/sidebar'
+import { Sidebar, useSidebar } from '@/features/sidebar'
 import { parse } from '@/helpers/parse'
 import { CONFIGURATION_SYMBOL } from '@/hooks/useConfig'
 import { useNavState } from '@/hooks/useNavState'
@@ -458,12 +457,12 @@ watch(hash, (newHash, oldHash) => {
               <ClassicHeader v-if="configuration.layout === 'classic'">
                 <div
                   v-if="$slots['document-selector']"
-                  class="w-64 empty:hidden">
+                  class="w-64 *:!p-0 empty:hidden">
                   <slot name="document-selector" />
                 </div>
                 <SearchButton
                   v-if="!configuration.hideSearch"
-                  class="t-doc__sidebar"
+                  class="t-doc__sidebar max-w-64"
                   :searchHotKey="configuration.searchHotKey"
                   :spec="parsedDocument" />
                 <template #dark-mode-toggle>
@@ -702,9 +701,9 @@ watch(hash, (newHash, oldHash) => {
 }
 </style>
 <style scoped>
-/** 
-* Sidebar CSS for standalone 
-* TODO: @brynn move this to the sidebar block OR the ApiReferenceStandalone component 
+/**
+* Sidebar CSS for standalone
+* TODO: @brynn move this to the sidebar block OR the ApiReferenceStandalone component
 * when the new elements are available
 */
 @media (max-width: 1000px) {