+++ packages/api-reference/src/features/sidebar/helpers/create-sidebar.test.ts
@@ -1,9 +1,9 @@
-import type { OpenAPIV3_1 } from '@scalar/openapi-types'
-import { describe, expect, it, vi } from 'vitest'
-import { toRef, toValue, ref } from 'vue'
-import { createSidebar } from './create-sidebar'
-import { apiReferenceConfigurationSchema } from '@scalar/types/api-reference'
 import { useNavState } from '@/hooks/useNavState'
+import type { OpenAPIV3_1 } from '@scalar/openapi-types'
+import { apiReferenceConfigurationSchema } from '@scalar/types/api-reference'
+import { describe, expect, it, vi } from 'vitest'
+import { ref, toRef, toValue } from 'vue'
+import { createSidebar } from './create-sidebar'
 
 // Mock vue's inject
 vi.mock('vue', () => {
@@ -1208,6 +1208,30 @@ describe('createSidebar', () => {
             ],
           },
         ],
+      })
+    })
+
+    it('uses the title attribute of the schema', () => {
+      expect(
+        createSidebar(
+          ref({
+            openapi: '3.1.0',
+            info: {
+              title: 'Hello World',
+              version: '1.0.0',
+            },
+            components: {
+              schemas: {
+                Planet: {
+                  title: 'Foobar',
+                },
+              },
+            },
+          } as OpenAPIV3_1.Document),
+          mockOptions,
+        ).items.value,
+      ).toMatchObject({
+        entries: [{ title: 'Models', children: [{ title: 'Foobar' }] }],
       })
     })