+++ packages/components/src/components/ScalarTooltip/index.ts
@@ -1,1 +1,4 @@
 export { default as ScalarTooltip } from './ScalarTooltip.vue'
+
+export type { ScalarTooltipPlacement, TooltipConfiguration } from './types'
+export { useTooltip } from './useTooltip'