+++ integrations/sveltekit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scalar/sveltekit
 
+## 0.1.10
+
+### Patch Changes
+
+- @scalar/core@0.3.4
+
 ## 0.1.9
 
 ### Patch Changes